ref(): moved wal2json messages to messages package to fix visibility issue
