+++ .github/workflows/go.yml
@@ -15,7 +15,7 @@ jobs:
     runs-on: ubuntu-latest
     strategy:
       matrix:
-        go-version: [ "1.19", "1.20" ]
+        go-version: [ "1.21.1" ]
         pg-version: [ 13, 14, 15 ]
       fail-fast: false
     env:
@@ -26,15 +26,7 @@ jobs:
     - name: Set up Go
       uses: actions/setup-go@v4
       with:
-        go-version: '1.20'
+        go-version: '1.21.1'
 
     - name: Build
       run: go build -v ./...
-
-    - name: Stand up Postgres ${{ matrix.pg-version }}
-      run: docker-compose up -d postgres
-      env:
-        POSTGRES_VERSION: ${{ matrix.pg-version }}
-
-    - name: Test
-      run: go test -v ./...
