+++ internal/replication/snapshotter.go
@@ -4,25 +4,12 @@ import (
 	"context"
 	"database/sql"
 	"fmt"
-	"github.com/apache/arrow/go/v14/arrow"
 	"github.com/jackc/pgx/v5"
 	"github.com/jackc/pgx/v5/pgconn"
 	_ "github.com/lib/pq"
 	"log"
 	"strings"
 )
-
-type Wal2JsonChanges struct {
-	Lsn     string
-	Changes []Wal2JsonChange `json:"change"`
-}
-
-type Wal2JsonChange struct {
-	Kind   string       `json:"action"`
-	Schema string       `json:"schema"`
-	Table  string       `json:"table"`
-	Row    arrow.Record `json:"data"`
-}
 
 type Snapshotter struct {
 	pgConnection *pgx.Conn
